[FEAT] PhotoTopic: PhotoTopicView 분리 #1
